requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiolimiter==1.2.1
aiosignal==1.3.2
alembic==1.13.1
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
argcomplete==3.6.2
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
boto3==1.38.40
boto3-stubs==1.38.40
botocore==1.38.40
botocore-stubs==1.38.30
bytecode==0.16.2
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
cohere==5.15.0
colorama==0.4.6
datadog==0.51.0
datadog_lambda==6.110.0
ddtrace==3.9.3
dill==0.4.0
distro==1.9.0
docling==2.34.0
docling-core==2.38.0
docling-ibm-models==3.5.0
docling-parse==4.0.5
docstring_parser==0.16
easyocr==1.7.2
einops==0.8.1
envier==0.6.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fasta2a==0.3.1
fastapi==0.115.13
fastavro==1.11.1
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.5.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.97.0
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.21.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
griffe==1.7.3
groq==0.28.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
hf-xet==1.1.4
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
idna==3.10
ijson==3.4.0
imageio==2.37.0
importlib_metadata==8.7.0
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain-core==0.3.65
langchain-openai==0.3.24
langchain-text-splitters==0.3.8
langsmith==0.3.45
latex2mathml==3.78.0
lazy_loader==0.4
litellm==1.72.7
logfire-api==3.21.1
lxml==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
marko==2.1.4
MarkupSafe==3.0.2
mcp==1.9.4
mdurl==0.1.2
mistral_common==1.6.2
mistralai==1.8.2
mpire==2.10.2
mpmath==1.3.0
multidict==6.5.0
multiprocess==0.70.18
mypy-boto3-s3==1.38.39
mypy-boto3-sts==1.38.38
networkx==3.5
ninja==1.11.1.4
numpy==2.3.0
ollama==0.5.1
openai==1.88.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.34.1
orjson==3.10.18
oso-cloud==2.5.0
packaging==24.2
pandas==2.3.0
pgai==0.11.2
pgvector==0.4.1
pillow==11.2.1
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic-ai==0.3.1
pydantic-ai-slim==0.3.1
pydantic-evals==0.3.1
pydantic-graph==0.3.1
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pylatexenc==2.10
PyMuPDF==1.26.1
pymupdf4llm==0.0.24
pypdfium2==4.30.1
pytest==8.4.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-multipart==0.0.20
python-pptx==1.0.2
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
rtree==1.4.0
s3transfer==0.13.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
semchunk==2.2.2
semver==3.0.4
sentence-transformers==4.1.0
sentencepiece==0.2.0
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.3.6
starlette==0.46.2
structlog==25.4.0
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.6.0
tifffile==2025.6.11
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.4
typer==0.15.4
types-awscrt==0.27.2
types-requests==2.32.4.20250611
types-s3transfer==0.13.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.34.3
voyageai==0.3.1
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
xlsxwriter==3.2.5
xmltodict==0.14.2
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
